Set additional email and phone properties as per foaf spec + refactoring
