+++ lib/session.js
@@ -1,16 +1,11 @@
 import { uuid, sparqlEscapeUri, sparqlEscapeString, sparqlEscapeDateTime } from 'mu';
 import { querySudo as query, updateSudo as update } from './auth-sudo';
-
-const allowNoRoleClaim = process.env.MU_APPLICATION_AUTH_ALLOW_NO_ROLE_CLAIM === 'true';
-const defaultGroupUri = process.env.MU_APPLICATION_AUTH_DEFAULT_GROUP_URI;
+import { ensureUser } from './user';
 
 const serviceHomepage = 'https://github.com/lblod/acmidm-login-service';
 const resourceBaseUri = process.env.MU_APPLICATION_RESOURCE_BASE_URI || 'http://data.lblod.info/';
-const personResourceBaseUri = `${resourceBaseUri}id/persoon/`;
 const accountResourceBaseUri = `${resourceBaseUri}id/account/`;
-const identifierResourceBaseUri = `${resourceBaseUri}id/identificator/`;
 
-const userIdClaim = process.env.MU_APPLICATION_AUTH_USERID_CLAIM || 'rrn';
 const accountIdClaim = process.env.MU_APPLICATION_AUTH_ACCOUNTID_CLAIM || 'vo_id';
 
 const removeOldSessions = async function (sessionUri) {
@@ -37,153 +32,11 @@ const removeCurrentSession = async function (sessionUri) {
 };
 
 const ensureUserAndAccount = async function (claims, groupId) {
-  const graph = `http://mu.semte.ch/graphs/public`;
+  const graph = 'http://mu.semte.ch/graphs/public';
   const { personUri } = await ensureUser(claims, graph);
   const { accountUri, accountId } = await ensureAccountForUser(personUri, claims, graph);
   return { accountUri, accountId };
 };
-
-const ensureUser = async function (claims, graph) {
-  const userId = claims[userIdClaim];
-
-  const queryResult = await query(`
-    PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-    PREFIX adms: <http://www.w3.org/ns/adms#>
-    PREFIX dcterms: <http://purl.org/dc/terms/>
-
-    SELECT ?person ?personId
-    FROM <${graph}> {
-      ?person a foaf:Person ;
-            mu:uuid ?personId ;
-            adms:identifier ?identifier .
-      ?identifier skos:notation ${sparqlEscapeString(userId)} .
-    }`);
-
-  if (queryResult.results.bindings.length) {
-    const result = queryResult.results.bindings[0];
-    const personUri = result.person.value;
-    await updateUserNames(personUri, claims, graph);
-    return { personUri: personUri, personId: result.personId.value };
-  } else {
-    const { personUri, personId } = await insertNewUser(claims, graph);
-    return { personUri, personId };
-  }
-};
-
-const updateUserNames = async function (personUri, claims, graph) {
-  let updateQuery = `
-    PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-    PREFIX adms: <http://www.w3.org/ns/adms#>
-    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-
-    DELETE {
-      GRAPH <${graph}> {
-        ${sparqlEscapeUri(personUri)} foaf:firstName ?first ;
-                                      foaf:familyName ?last .
-      }
-    } WHERE {
-      GRAPH <${graph}> {
-        ${sparqlEscapeUri(personUri)} a foaf:Person .
-        OPTIONAL {
-          ${sparqlEscapeUri(personUri)} foaf:firstName ?first .
-        }
-        OPTIONAL {
-          ${sparqlEscapeUri(personUri)} foaf:familyName ?last .
-        }
-      }
-    }`;
-
-  await update(updateQuery);
-
-  updateQuery = `
-  PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-  PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-  PREFIX adms: <http://www.w3.org/ns/adms#>
-  PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-
-  INSERT DATA {
-    GRAPH <${graph}> {
-      ${sparqlEscapeUri(personUri)} a foaf:Person .
-  `;
-
-  if (claims.given_name) {
-    updateQuery += `${sparqlEscapeUri(personUri)} foaf:firstName ${sparqlEscapeString(claims.given_name)} . `;
-  }
-  if (claims.family_name) {
-    updateQuery += `${sparqlEscapeUri(personUri)} foaf:familyName ${sparqlEscapeString(claims.family_name)} . `;
-  }
-  updateQuery += `
-      }
-    }
-  `;
-
-  await update(updateQuery);
-};
-
-const insertNewUser = async function (claims, graph) {
-  const personId = uuid();
-  const person = `${personResourceBaseUri}${personId}`;
-  const identifierId = uuid();
-  const identifier = `${identifierResourceBaseUri}${identifierId}`;
-
-  let insertData = `
-    PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-    PREFIX adms: <http://www.w3.org/ns/adms#>
-    PREFIX skos: <http://www.w3.org/2004/02/skos/core#>
-
-    INSERT DATA {
-      GRAPH <${graph}> {
-        ${sparqlEscapeUri(person)} a foaf:Person ;
-                                 mu:uuid ${sparqlEscapeString(personId)} ;
-                                 adms:identifier ${sparqlEscapeUri(identifier)} .
-        ${sparqlEscapeUri(defaultGroupUri)} foaf:member ${sparqlEscapeUri(person)} .
-        ${sparqlEscapeUri(identifier)} a adms:Identifier ;
-                                       mu:uuid ${sparqlEscapeString(identifierId)} ;
-                                       skos:notation ${sparqlEscapeString(claims[userIdClaim])} .
-    `;
-
-  if (claims.given_name) {
-    insertData += `${sparqlEscapeUri(person)} foaf:firstName ${sparqlEscapeString(claims.given_name)} . `;
-  }
-  if (claims.family_name) {
-    insertData += `${sparqlEscapeUri(person)} foaf:familyName ${sparqlEscapeString(claims.family_name)} . `;
-  }
-  insertData += `
-      }
-    }
-  `;
-
-  await update(insertData);
-
-  return { personUri: person, personId: personId };
-};
-
-const checkIfEmailShouldBeUpdated = function (result, claims) {
-  if(!result) {
-    return false;
-  }
-
-  if(result.email) {
-    return result.email.value === claims.vo_email;
-  }
-
-  return false;
-};
-
-const checkIfPhoneShouldbeUpdated = function(result, claims) {
-  if(!result) {
-    return false;
-  }
-
-  if(result.phone) {
-    return result.phone.value === claims.phone;
-  }
-
-  return false;
-}
 
 const ensureAccountForUser = async function (personUri, claims, graph) {
   const accountId = claims[accountIdClaim];
@@ -196,82 +49,20 @@ const ensureAccountForUser = async function (personUri, claims, graph) {
 
     SELECT ?account ?accountId ?email ?phone
     FROM <${graph}> {
-      ${sparqlEscapeUri(personUri)} foaf:account ?account .
-      ?account a foaf:OnlineAccount ;
-               mu:uuid ?accountId ;
-               dcterms:identifier ${sparqlEscapeString(accountId)} .
-              OPTIONAL {
-                 ?account ext:email ?email
-              }
-              OPTIONAL {
-                 ?account ext:phone ?phone
-              }
+        ${sparqlEscapeUri(personUri)} foaf:account ?account .
+        ?account a foaf:OnlineAccount ;
+            mu:uuid ?accountId ;
+            dcterms:identifier ${sparqlEscapeString(accountId)} .
     }`);
-
-  //TODO: Optional om te checken of email reeds ingevuld is of verschillend is van de email in claims
 
   if (queryResult.results.bindings.length) {
     const result = queryResult.results.bindings[0];
-
-    if(checkIfEmailShouldBeUpdated(result,claims)) {
-      await updateUserAccount(result.account.value,claims,graph);
-    }
-
-    if(checkIfPhoneShouldbeUpdated(result,claims)) {
-      await updateUserAccount(result.account.value, claims,graph);
-    }
-
     return { accountUri: result.account.value, accountId: result.accountId.value };
   } else {
     const { accountUri, accountId } = await insertNewAccountForUser(personUri, claims, graph);
     return { accountUri, accountId };
   }
 };
-
-
-const updateUserAccount = async function (accountUri, claims, graph) {
-  let updateQuery = `
-    PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-    PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
-
-    DELETE {
-      GRAPH <${graph}> {
-        ${sparqlEscapeUri(accountUri)}  ext:email ?email .
-      }
-    } WHERE {
-      GRAPH <${graph}> {
-        ${sparqlEscapeUri(accountUri)} a foaf:OnlineAccount ;
-        OPTIONAL {
-          ${sparqlEscapeUri(accountUri)} mu:uuid ?accountId .
-        }
-        OPTIONAL {
-          ${sparqlEscapeUri(accountUri)} ext:email ?email .
-        }
-         OPTIONAL {
-          ${sparqlEscapeUri(accountUri)} ext:phone ?phone .
-        }
-      }
-    }`;
-
-  await update(updateQuery);
-
-  updateQuery = `
-  PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
-
-  INSERT DATA {
-    GRAPH <${graph}> {
-      ${sparqlEscapeUri(accountUri)} ext:email ${sparqlEscapeString(claims.vo_email)} ;
-                                     ext:phone ${sparqlEscapeString(claims.phone)} .
-  `;
-  updateQuery += `
-      }
-    }
-  `;
-
-  await update(updateQuery);
-};
-
 
 const insertNewAccountForUser = async function (person, claims, graph) {
   const accountId = uuid();
@@ -299,12 +90,6 @@ const insertNewAccountForUser = async function (person, claims, graph) {
   }
   if (claims.vo_doelgroepnaam) {
     insertData += `${sparqlEscapeUri(account)} acmidm:doelgroepNaam ${sparqlEscapeString(claims.vo_doelgroepnaam)} . `;
-  }
-  if (claims.vo_email) {
-    insertData += `${sparqlEscapeUri(account)} ext:email ${sparqlEscapeString(claims.vo_email)} . `;
-  }
-  if (claims.phone) {
-    insertData += `${sparqlEscapeUri(account)} ext:phone ${sparqlEscapeString(claims.vo_email)} . `;
   }
 
   insertData += `
@@ -344,34 +129,6 @@ const insertNewSessionForAccount = async function (accountUri, sessionUri, group
 
   await update(insertData);
   return { sessionUri, sessionId };
-};
-
-const selectUserGroup = async function (accountUri, claims, roleClaim) {
-  if ((claims[roleClaim] && claims[roleClaim].length > 0) ||
-    allowNoRoleClaim) {
-    const queryResult = await query(`
-    PREFIX mu: <http://mu.semte.ch/vocabularies/core/>
-    PREFIX ext: <http://mu.semte.ch/vocabularies/ext/>
-    PREFIX session: <http://mu.semte.ch/vocabularies/session/>
-    PREFIX foaf: <http://xmlns.com/foaf/0.1/>
-    PREFIX besluit: <http://data.vlaanderen.be/ns/besluit#>
-
-    SELECT ?group ?groupId ?groupName
-    WHERE {
-      GRAPH <http://mu.semte.ch/graphs/public> {
-          ?user foaf:account <${accountUri}> .
-          ?group foaf:member ?user.
-          ?group foaf:name ?groupName.
-          ?group mu:uuid ?groupId.
-      }
-    }`);
-
-    if (queryResult.results.bindings.length) {
-      const result = queryResult.results.bindings[0];
-      return { groupUri: result.group.value, groupId: result.groupId.value, groupName: result.groupName.value };
-    }
-  }
-  return { groupUri: null, groupId: null, groupName: null };
 };
 
 const selectAccountBySession = async function (session) {
@@ -445,8 +202,6 @@ export {
   removeCurrentSession,
   ensureUserAndAccount,
   insertNewSessionForAccount,
-  selectUserGroup,
   selectAccountBySession,
-  selectCurrentSession,
-  updateUserAccount
+  selectCurrentSession
 };
